refactor(test): Correct import path for `utils` module
